refactor: extract shared weekly report helpers into weekly-report-common.mjs
The slow-tests report imported plumbing out of the flaky report, which
coupled the two scripts and forced the slow workflow to sparse-checkout
its sibling. Move the shared machinery - request/retry, HogQL, repo-path
and owner resolution, Slack table cells, the Slack post, and the
edit-this-workflow footer - into weekly-report-common.mjs and point both
reports and both workflow sparse checkouts at it.

postToSlack's notification text is now a required argument instead of a
flaky-flavoured default. Behavior is otherwise unchanged; the moved
functions keep their existing test coverage (slow 10, flaky 14).

Generated-By: PostHog Desktop
Task-Id: 9fd4178c-80ed-4c47-acee-d903f5808186
