+++ .github/scripts/weekly-flaky-report.mjs
@@ -14,24 +14,33 @@
 // retry-enabled lanes. Master-burst breakage and branch-only tests are filtered
 // out client-side.
 
-import { execFileSync } from 'node:child_process'
 import { pathToFileURL } from 'node:url'
 
-const HOST = (process.env.POSTHOG_HOST || 'https://us.posthog.com').replace(/\/$/, '')
-const PROJECT_ID = process.env.POSTHOG_PROJECT_ID || ''
-const API_KEY = process.env.POSTHOG_API_KEY || ''
+import {
+    AUTH_HEADERS,
+    cell,
+    DRY_RUN,
+    editWorkflowBlock,
+    GITHUB_REPOSITORY,
+    GITHUB_SERVER_URL,
+    hogql,
+    HOST,
+    linkedCell,
+    postToSlack,
+    PROJECT_ID,
+    API_KEY,
+    repoPathResolver,
+    requestPosthog,
+    resolveOwners,
+    shortName,
+    SLACK_BOT_TOKEN,
+    SLACK_CHANNEL,
+} from './weekly-report-common.mjs'
+
 const SOURCE_ID = process.env.ENG_ANALYTICS_SOURCE_ID || ''
 // The synced runs table name carries the warehouse source prefix, which differs per project.
 const RUNS_TABLE = process.env.ENG_ANALYTICS_RUNS_TABLE || 'eng_analyticsgithub_workflow_runs'
 const TRUNK_TABLE = process.env.TRUNK_QUARANTINE_TABLE || 'trunkio.quarantinedtests'
-const SLACK_BOT_TOKEN = process.env.SLACK_BOT_TOKEN || ''
-const SLACK_CHANNEL = process.env.SLACK_CHANNEL || 'C09ADEV3AJD' // #flakey-tests
-const DRY_RUN = ['1', 'true', 'yes'].includes((process.env.DRY_RUN || '').toLowerCase())
-
-const GITHUB_SERVER_URL = process.env.GITHUB_SERVER_URL || 'https://github.com'
-const GITHUB_REPOSITORY = process.env.GITHUB_REPOSITORY || 'PostHog/posthog'
-const GITHUB_WORKFLOW_REF = process.env.GITHUB_WORKFLOW_REF || ''
-const GITHUB_REF_NAME = process.env.GITHUB_REF_NAME || 'master'
 
 const TOP_N = 10
 const CANDIDATE_POOL = 40
@@ -48,41 +57,6 @@ const RUNNER_LABELS = { pytest: 'pytest', jest: 'Jest' }
 const TRUNK_UPLOADS_ON = process.env.TRUNK_UPLOAD_ENABLED === 'true'
 const TRUNK_MASKS_CI = TRUNK_UPLOADS_ON && process.env.TRUNK_QUARANTINE_ENABLED === 'true'
 
-const RETRY_ATTEMPTS = 3
-const RETRY_DELAY_MS = 30_000
-
-async function request(url, options, label) {
-    const res = await fetch(url, { ...options, signal: AbortSignal.timeout(150_000) })
-    const body = await res.text()
-    const fail = (detail, retryable) => {
-        throw Object.assign(new Error(`${label} -> ${res.status}: ${detail}`), { retryable })
-    }
-    let parsed
-    try {
-        parsed = JSON.parse(body)
-    } catch {
-        fail(`non-JSON response (${body.slice(0, 120)})`, true)
-    }
-    if (!res.ok) {
-        fail(parsed?.detail || body, res.status >= 500 || res.status === 429)
-    }
-    return parsed
-}
-
-async function withRetry(fn) {
-    for (let attempt = 1; ; attempt++) {
-        try {
-            return await fn()
-        } catch (err) {
-            if (err.retryable === false || attempt >= RETRY_ATTEMPTS) {
-                throw err
-            }
-            console.warn(`${err.message} — attempt ${attempt}/${RETRY_ATTEMPTS}, retrying in ${RETRY_DELAY_MS / 1000}s`)
-            await new Promise((resolve) => setTimeout(resolve, RETRY_DELAY_MS))
-        }
-    }
-}
-
 function endpointUrl(action, params = {}) {
     const url = new URL(`${HOST}/api/projects/${PROJECT_ID}/engineering_analytics/${action}/`)
     for (const [k, v] of Object.entries(params)) {
@@ -96,8 +70,6 @@ function endpointUrl(action, params = {}) {
     return url
 }
 
-const AUTH_HEADERS = { Authorization: `Bearer ${API_KEY}` }
-
 function flakyTestsUrl(runner) {
     return endpointUrl('flaky_tests', {
         date_from: '-7d',
@@ -108,23 +80,7 @@ function flakyTestsUrl(runner) {
 }
 
 function fetchFlakyTests(runner) {
-    return withRetry(() => request(flakyTestsUrl(runner), { headers: AUTH_HEADERS }, 'flaky_tests'))
-}
-
-// `values` bind through HogQL's {placeholder} syntax, escaped server-side — never
-// concatenate attacker-controlled test ids into the query source.
-function hogql(query, values) {
-    return withRetry(() =>
-        request(
-            `${HOST}/api/projects/${PROJECT_ID}/query/`,
-            {
-                method: 'POST',
-                headers: { ...AUTH_HEADERS, 'Content-Type': 'application/json' },
-                body: JSON.stringify({ query: { kind: 'HogQLQuery', query, values } }),
-            },
-            'query'
-        )
-    )
+    return requestPosthog(flakyTestsUrl(runner), { headers: AUTH_HEADERS }, 'flaky_tests')
 }
 
 // A same-commit recovery proves a flake, so only suppress likely one-merge bursts
@@ -169,69 +125,6 @@ async function fetchCandidatePools(runners, toRepoPaths, fetchTests = fetchFlaky
             return { runner, candidates: selectReportCandidates(result.items || [], runner, toRepoPaths) }
         })
     )
-}
-
-// Product suites run from their product dir, so a selector path may be repo- or
-// product-relative — suffix-match the tracked index (full even under sparse checkout).
-function trackedTestPaths(runGit = execFileSync) {
-    // The tracked-file list is a few MB; the 1MB execFileSync default truncates it.
-    return runGit('git', ['ls-files', '*.py', '*.js', '*.jsx', '*.ts', '*.tsx'], {
-        encoding: 'utf8',
-        maxBuffer: 64 * 1024 * 1024,
-    })
-        .split('\n')
-        .filter(Boolean)
-}
-
-function repoPathResolver(trackedPaths = trackedTestPaths()) {
-    const trackedSet = new Set(trackedPaths)
-    const bySuffix = new Map()
-    for (const path of trackedPaths) {
-        const base = path.split('/').pop()
-        if (!bySuffix.has(base)) {
-            bySuffix.set(base, [])
-        }
-        bySuffix.get(base).push(path)
-    }
-    return (selectorPath) => {
-        if (trackedSet.has(selectorPath)) {
-            return [selectorPath]
-        }
-        return (bySuffix.get(selectorPath.split('/').pop()) || []).filter((p) => p.endsWith(`/${selectorPath}`))
-    }
-}
-
-// Ambiguous suffix matches only count when every candidate agrees on the owner.
-function resolveOwners(items, toRepoPaths = repoPathResolver()) {
-    const candidates = new Map()
-    for (const item of items) {
-        const selectorPath = item.selector.split('::')[0]
-        if (!candidates.has(selectorPath)) {
-            candidates.set(selectorPath, toRepoPaths(selectorPath))
-        }
-    }
-    const allPaths = [...new Set([...candidates.values()].flat())]
-    let resolved = {}
-    if (allPaths.length > 0) {
-        try {
-            const out = execFileSync('python3', ['-m', 'posthog_owners'], {
-                encoding: 'utf8',
-                input: allPaths.join('\n'),
-                env: { ...process.env, PYTHONPATH: 'tools/owners' },
-            })
-            resolved = JSON.parse(out)
-        } catch (err) {
-            // Degrade to "unowned" rather than skipping the week's post.
-            console.warn(`owners resolution failed — reporting all tests as unowned: ${err.message}`)
-        }
-    }
-    return (item) => {
-        const selectorPath = item.selector.split('::')[0]
-        const paths = candidates.get(selectorPath) || []
-        const owners = new Set(paths.map((p) => (resolved[p]?.owners || [])[0] || 'unowned'))
-        const owner = owners.size === 1 ? [...owners][0] : 'unowned'
-        return { owner, repoPath: paths.length === 1 ? paths[0] : null }
-    }
 }
 
 // One test carries different leading path segments depending on who named it: a product suite
@@ -457,23 +350,6 @@ async function buildRunnerReports(
     )
 }
 
-function cell(text) {
-    return { type: 'raw_text', text }
-}
-
-function linkedCell(links) {
-    const elements = links.flatMap(({ url, text }, index) => [
-        ...(index > 0 ? [{ type: 'text', text: ' ' }] : []),
-        { type: 'link', url, text },
-    ])
-    return { type: 'rich_text', elements: [{ type: 'rich_text_section', elements }] }
-}
-
-function shortName(selector) {
-    const name = selector.split('::').pop()
-    return name.length > 36 ? `${name.slice(0, 35)}…` : name
-}
-
 function tableRows(items, ownerFor, extrasFor, statusFor = () => null) {
     return items.map((item) => {
         const { owner, repoPath } = ownerFor(item)
@@ -537,32 +413,11 @@ function buildBlocks(now, rows) {
             ],
         },
     ]
-    const workflowPath = GITHUB_WORKFLOW_REF.split('@')[0].replace(`${GITHUB_REPOSITORY}/`, '')
-    if (GITHUB_REPOSITORY && workflowPath) {
-        const editUrl = `${GITHUB_SERVER_URL}/${GITHUB_REPOSITORY}/edit/${GITHUB_REF_NAME}/${workflowPath}`
-        blocks.push({ type: 'context', elements: [{ type: 'mrkdwn', text: `<${editUrl}|edit this workflow>` }] })
+    const editBlock = editWorkflowBlock()
+    if (editBlock) {
+        blocks.push(editBlock)
     }
     return blocks
-}
-
-async function postToSlack(blocks, text = 'Weekly flaky test report') {
-    const res = await fetch('https://slack.com/api/chat.postMessage', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json; charset=utf-8', Authorization: `Bearer ${SLACK_BOT_TOKEN}` },
-        body: JSON.stringify({
-            channel: SLACK_CHANNEL,
-            blocks,
-            text, // notification fallback
-            unfurl_links: false,
-        }),
-    })
-    const data = await res.json()
-    if (!data.ok) {
-        const validationDetails = Array.isArray(data.response_metadata?.messages)
-            ? `: ${data.response_metadata.messages.join('; ')}`
-            : ''
-        throw new Error(`Slack chat.postMessage failed: ${data.error}${validationDetails}`)
-    }
 }
 
 async function main() {
@@ -591,7 +446,7 @@ async function main() {
     if (!SLACK_BOT_TOKEN) {
         throw new Error('SLACK_BOT_TOKEN not set on a non-dry run — refusing to silently skip.')
     }
-    await postToSlack(blocks)
+    await postToSlack(blocks, 'Weekly flaky test report')
     console.info(`Posted weekly flaky report to ${SLACK_CHANNEL}.`)
 }
 
@@ -606,22 +461,14 @@ export {
     buildBlocks,
     buildQueue,
     buildRunnerReports,
-    cell,
     CLUSTER_MIN_TESTS,
     enrich,
     enrichRunnerCandidates,
     fetchCandidatePools,
     fetchTrunkQuarantined,
     flakyTestsUrl,
-    hogql,
-    linkedCell,
-    postToSlack,
     REPORT_RUNNERS,
-    repoPathResolver,
-    resolveOwners,
     selectReportCandidates,
-    shortName,
     tableRows,
     testStatusFor,
-    trackedTestPaths,
 }
