+++ .github/workflows/weekly-slow-tests-report.yml
@@ -46,7 +46,7 @@ jobs:
               with:
                   sparse-checkout: |
                       .github/scripts/weekly-slow-tests-report.mjs
-                      .github/scripts/weekly-flaky-report.mjs
+                      .github/scripts/weekly-report-common.mjs
                       .nvmrc
                       tools/owners
                       owners.yaml
